- refactor pointsHistory renderable
